VIDIX: fix variable type to match bytes read.
git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@37726 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ vidix/sysdep/pci_linux.c
@@ -192,7 +192,7 @@ static long pci_config_read_long(
           int func,
           unsigned cmd)
 {
-    long retval;
+    unsigned retval;
     char path[100];
     int fd;
     sprintf(path,"/proc/bus/pci/%02d/%02x.0", bus, dev);
